+++ src/Compute/Compute/ChangeLog.md
@@ -20,6 +20,15 @@
 
 -->
 ## Upcoming Release
+* Added `-Hibernate` switch parameter to `Stop-AzVmss` default parameter set. 
+* For `Get-AzVmRunCommand`, a bug is fixed to work when returning a list of RunCommands [#22403]
+* Updated Azure.Core to 1.34.0.
+* Added the `Standard` value to the `SecurityType` parameter to the cmdlets `Set-AzDiskSecurityType`, `New-AzvmssConfig`, `Set-AzVmssSecurityProfile`, `Update-AzVmss`, `New-AzVmss`, `New-AzVMConfig`, `Set-AzVMsecurityProfile`, and `New-AzVM`.
+
+## Version 6.2.0
+* Fixed the `Update-AzVmss` cmdlet so the `AutomaticRepairGracePeriod`, `AutomaticRepairAction`, and `EnableAutomaticRepair` parameters function correctly.
+* Updated help doc for `New-AzVM`, `New-AzVMConfig`, `New-AzVmss`, `New-AzVmssConfig`, `Update-AzVM`, and `Update-AzVmss` to include parameters that were previously added for Trusted Launch features.
+* Updated Azure.Core to 1.33.0.
 
 ## Version 6.1.0
 * Added useful examples to the `New-AzVMConfig` help doc.
@@ -29,7 +38,7 @@
 * Added a message to the user when they provide an outdated image alias to `New-AzVM` via the `-Image` parameter or to `New-AzVmss` via the `-ImageName` parameter.
   The non-versioned image aliases were updated to versioned values in October 2023, and this message is to help urge customers to use the newer versioned image alias values.
 * Changed the installation behavior for the `GuestAttestation` extension in `New-AzVM` and `New-AzVmss` to set the property `EnableAutomaticUpgrade` to true.
-* Changed to `Set-AzVMOperatingSystem` to correct unnecessary mandatory parameters.
+* Changes to `Set-AzVMOperatingSystem` to correct unnecessary mandatory parameters.
 * Changed the region for example 3 to eastus2 in the `New-AzVM` doc to resolve SKU not available errors.
 
 ## Version 6.0.0